fix ALL

+++ trainer/trainer.py
@@ -40,7 +40,7 @@ class Trainer(BaseTrainer):
         self.model.train()
         self.train_metrics.reset()
         for batch_idx, (data, target) in enumerate(self.data_loader):
-            data, target = data.to(self.device), target.to(self.device)
+            data, target = data.to(self.device), target.to(self.device).long()
 
             self.optimizer.zero_grad()
             output = self.model(data)
@@ -83,7 +83,7 @@ class Trainer(BaseTrainer):
         self.valid_metrics.reset()
         with torch.no_grad():
             for batch_idx, (data, target) in enumerate(self.valid_data_loader):
-                data, target = data.to(self.device), target.to(self.device)
+                data, target = data.to(self.device), target.to(self.device).long()
 
                 output = self.model(data)
                 loss = self.criterion(output, target)
